Major change to how aligning works
I didn't like how labels or derectives would have large spacing between
the words.

I looked up an examples of assembly code and tried to mimic the spacing
to the best of my ability.

As of this patch, I have yet to test it on a large file but I imagine
more changes will come in the future.